+++ src/utils/productModifiers.js
@@ -8,7 +8,6 @@ const getProductName = productName => {
 };
 
 const getSneakersSizes = product => {
-  console.log(product);
   const prod = product.constructor === Array ? product : product._sub_products;
   if (prod.length > 0) {
     return prod.map(subProd => {
